[dali] tweak color temperature abs QuantityType fix (#14029)
adds an example of usage of the channel to the README, and
simplifies one method call

+++ bundles/org.openhab.binding.dali/src/main/java/org/openhab/binding/dali/internal/handler/DaliDt8DeviceHandler.java
@@ -76,7 +76,7 @@ public class DaliDt8DeviceHandler extends DaliDeviceHandler {
                         logger.warn("Unable to convert command {} to mireks", command);
                         return;
                     }
-                    mirek = commandQuantity.toBigDecimal().intValue();
+                    mirek = commandQuantity.intValue();
                 } else {
                     logger.warn("Unable to convert command {} to mireks", command);
                     return;
